chore: Refactor and simplify builders
