+++ packages/core/src/data/formatters.ts
@@ -1,6 +1,3 @@
-import { Result } from 'sarif';
-import { sumOccurrences } from '../utils/sarif-utils';
-
 export function toPercent(numeratorOrValue: number, denominator?: number): string {
   let value: number =
     typeof denominator === 'number' ? numeratorOrValue / denominator : numeratorOrValue;
@@ -14,18 +11,6 @@ export function groupDataByField<T>(results: T[], field: string): T[][] {
     map.get(getMultiLevelProp(result, field))?.push(result as never);
   });
   return [...map.values()];
-}
-
-/*
- * This function takes in multiple Results with different locations but the same message,
- * and return a single Result to be rendered in the console (with occurrenceCounts added together)
- */
-export function combineResultsForRendering(groupedResults: Result[][]): Result[] {
-  return groupedResults.map((results) => {
-    const result = results[0];
-    result.occurrenceCount = sumOccurrences(results);
-    return result;
-  });
 }
 
 /*
