+++ packages/core/src/data/path.ts
@@ -1,0 +1,7 @@
+export function normalizePath(path: string, cwd: string) {
+  return path.replace(`${cwd}/`, '');
+}
+
+export function normalizePaths(paths: string[], cwd: string) {
+  return paths.map((path) => normalizePath(path, cwd));
+}
